requirements.txt
pyspark==3.5.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2
xgboost==2.1.1
lightgbm==4.5.0
matplotlib==3.9.2
seaborn==0.13.2
pyarrow==17.0.0
tqdm==4.66.5
pydantic==2.8.2
kafka-python==2.0.2
joblib==1.4.2
requests==2.32.3
openpyxl==3.1.5
pypdf==4.3.1
jupyter==1.0.0
ipykernel==6.29.5
